@@ -18,8 +18,8 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.1.42+42
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
-  flutter: ^2.10.0
+  sdk: ">=2.12.0"
+  flutter: ^3.0.0
 
 dependencies:
 
